Separate validation functions into new file
